Remove duplication between the 2 ethiopian calendar versions
