+++ lib/orbf/rules_engine.rb
@@ -15,6 +15,7 @@ require_relative "./rules_engine/services/period_converter"
 require_relative "./rules_engine/services/period_iterator"
 require_relative "./rules_engine/services/ethiopian_period_converter"
 require_relative "./rules_engine/services/ethiopian_period_iterator"
+require_relative "./rules_engine/services/ethiopian_converter"
 require_relative "./rules_engine/services/ethiopian_calendar"
 require_relative "./rules_engine/services/ethiopian_v2_calendar"
 require_relative "./rules_engine/services/period_facts"
